project/requirements.txt
django==3.0.7
djangorestframework==3.11.0

djoser==2.0.3
djangorestframework-simplejwt==4.4.0